Refactor example package name, update gradle.
