+++ example/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:label="@string/app_name"
         android:theme="@style/CustomTheme"
         android:supportsRtl="true">
-        <activity android:name="com.amqtech.opensource.appintroexample.ui.MainActivity">
+        <activity android:name="com.github.appintro.example.ui.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -22,28 +22,28 @@
         </activity>
 
         <activity
-            android:name="com.amqtech.opensource.appintroexample.ui.mainTabs.intro.DefaultIntro"
+            android:name="com.github.appintro.example.ui.mainTabs.intro.DefaultIntro"
             android:label="@string/app_name"
             android:theme="@style/AppIntroTheme"/>
         <activity
-            android:name="com.amqtech.opensource.appintroexample.ui.mainTabs.intro.DefaultIntro2"
+            android:name="com.github.appintro.example.ui.mainTabs.intro.DefaultIntro2"
             android:label="@string/app_name"
             android:theme="@style/AppIntroTheme"/>
         <activity
-            android:name="com.amqtech.opensource.appintroexample.ui.mainTabs.intro.CustomLayoutIntro"
+            android:name="com.github.appintro.example.ui.mainTabs.intro.CustomLayoutIntro"
             android:label="@string/app_name"
             android:theme="@style/AppIntroTheme"/>
         <activity
-            android:name="com.amqtech.opensource.appintroexample.ui.mainTabs.intro.CustomBackgroundIntro"
+            android:name="com.github.appintro.example.ui.mainTabs.intro.CustomBackgroundIntro"
             android:label="@string/app_name"
             android:theme="@style/AppIntroTheme"/>
 
         <activity
-            android:name="com.amqtech.opensource.appintroexample.ui.permsTabs.intro.PermissionsIntro1"
+            android:name="com.github.appintro.example.ui.permsTabs.intro.PermissionsIntro1"
             android:label="@string/app_name"
             android:theme="@style/AppIntroTheme"/>
         <activity
-            android:name="com.amqtech.opensource.appintroexample.ui.permsTabs.intro.PermissionsIntro2"
+            android:name="com.github.appintro.example.ui.permsTabs.intro.PermissionsIntro2"
             android:label="@string/app_name"
             android:theme="@style/AppIntroTheme"/>
