+++ example/src/main/java/com/github/appintro/example/ui/mainTabs/DefaultLayout2Intro.java
@@ -1,4 +1,4 @@
-package com.amqtech.opensource.appintroexample.ui.mainTabs;
+package com.github.appintro.example.ui.mainTabs;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -9,21 +9,21 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 
-import com.amqtech.opensource.appintroexample.ui.mainTabs.intro.DefaultIntro;
+import com.github.appintro.example.ui.mainTabs.intro.DefaultIntro2;
 import com.github.paolorotolo.appintroexample.R;
 
-public class DefaultLayoutIntro extends Fragment {
+public class DefaultLayout2Intro extends Fragment {
 
     @SuppressWarnings("ConstantConditions")
     @Override
     public void onActivityCreated(@Nullable Bundle savedInstanceState) {
         super.onActivityCreated(savedInstanceState);
 
-        Button demo = getView().findViewById(R.id.intro1);
+        Button demo = getView().findViewById(R.id.intro2);
         demo.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                startActivity(new Intent(getActivity().getBaseContext(), DefaultIntro.class));
+                startActivity(new Intent(getActivity().getBaseContext(), DefaultIntro2.class));
             }
         });
     }
@@ -31,6 +31,6 @@ public class DefaultLayoutIntro extends Fragment {
     @Nullable
     @Override
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.tab_default_intro, container, false);
+        return inflater.inflate(R.layout.tab_default_intro2, container, false);
     }
 }
