+++ example/src/main/java/com/github/appintro/example/ui/permsTabs/PermissionsLayout1Intro.java
@@ -1,4 +1,4 @@
-package com.amqtech.opensource.appintroexample.ui.mainTabs;
+package com.github.appintro.example.ui.permsTabs;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -9,20 +9,21 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 
+import com.github.appintro.example.ui.permsTabs.intro.PermissionsIntro1;
 import com.github.paolorotolo.appintroexample.R;
 
-public class CustomLayoutIntro extends Fragment {
+public class PermissionsLayout1Intro extends Fragment {
 
     @SuppressWarnings("ConstantConditions")
     @Override
     public void onActivityCreated(@Nullable Bundle savedInstanceState) {
         super.onActivityCreated(savedInstanceState);
 
-        Button demo = getView().findViewById(R.id.intro3);
+        Button demo = getView().findViewById(R.id.intro1);
         demo.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                startActivity(new Intent(getActivity().getBaseContext(), com.amqtech.opensource.appintroexample.ui.mainTabs.intro.CustomLayoutIntro.class));
+                startActivity(new Intent(getActivity().getBaseContext(), PermissionsIntro1.class));
             }
         });
     }
@@ -30,6 +31,6 @@ public class CustomLayoutIntro extends Fragment {
     @Nullable
     @Override
     public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.tab_custom_layout_intro, container, false);
+        return inflater.inflate(R.layout.tab_permissions_1, container, false);
     }
 }
