+++ example/src/main/java/com/github/appintro/example/ui/permsTabs/intro/PermissionsIntro2.java
@@ -1,5 +1,6 @@
-package com.amqtech.opensource.appintroexample.ui.mainTabs.intro;
+package com.github.appintro.example.ui.permsTabs.intro;
 
+import android.Manifest;
 import android.graphics.Color;
 import android.os.Bundle;
 
@@ -11,7 +12,7 @@ import com.github.paolorotolo.appintro.AppIntroFragment;
 import com.github.paolorotolo.appintro.model.SliderPage;
 import com.github.paolorotolo.appintroexample.R;
 
-public class CustomBackgroundIntro extends AppIntro2 {
+public class PermissionsIntro2 extends AppIntro2 {
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
@@ -19,14 +20,14 @@ public class CustomBackgroundIntro extends AppIntro2 {
 
         SliderPage sliderPage1 = new SliderPage();
         sliderPage1.setTitle("Welcome!");
-        sliderPage1.setDescription("This is a demo of the AppIntro library, with a custom background on each slide!");
+        sliderPage1.setDescription("This is a demo of the AppIntro library, with permissions being requested on a slide!");
         sliderPage1.setImageDrawable(R.drawable.ic_slide1);
         sliderPage1.setBgColor(Color.TRANSPARENT);
         addSlide(AppIntroFragment.newInstance(sliderPage1));
 
         SliderPage sliderPage2 = new SliderPage();
-        sliderPage2.setTitle("Clean App Intros");
-        sliderPage2.setDescription("This library offers developers the ability to add clean app intros at the start of their apps.");
+        sliderPage2.setTitle("Permission Request");
+        sliderPage2.setDescription("In order to access your camera, you must give permissions.");
         sliderPage2.setImageDrawable(R.drawable.ic_slide2);
         sliderPage2.setBgColor(Color.TRANSPARENT);
         addSlide(AppIntroFragment.newInstance(sliderPage2));
@@ -45,8 +46,10 @@ public class CustomBackgroundIntro extends AppIntro2 {
         sliderPage4.setBgColor(Color.TRANSPARENT);
         addSlide(AppIntroFragment.newInstance(sliderPage4));
 
-        // Bind the background to the intro
-        setBackgroundResource(R.drawable.ic_drawer_header);
+        // Here we tell the library to request camera and location permission that are both required.
+        askForPermissions(new String[]{Manifest.permission.CAMERA, Manifest.permission.ACCESS_FINE_LOCATION}, 2, true);
+        // Here we request another permission that is instead not required.
+        askForPermissions(new String[]{Manifest.permission.WRITE_EXTERNAL_STORAGE}, 3, false);
     }
 
     @Override
