+++ example/src/main/java/com/github/appintro/example/util/TabLayout.java
@@ -1,4 +1,4 @@
-package com.amqtech.opensource.appintroexample.util;
+package com.github.appintro.example.util;
 
 import android.content.Context;
 import android.graphics.Typeface;
