+++ javadoc/overview-frame.html
@@ -18,7 +18,7 @@
         <li><a href="com/github/paolorotolo/appintroexample/package-frame.html"
                target="packageFrame">com.github.paolorotolo.appintroexample</a></li>
         <li><a href="com/github/paolorotolo/appintroexample/util/package-frame.html"
-               target="packageFrame">com.amqtech.opensource.appintroexample.util</a></li>
+               target="packageFrame">com.github.appintro.example.util</a></li>
     </ul>
 </div>
 <p>&nbsp;</p>
