+++ javadoc/overview-summary.html
@@ -87,7 +87,7 @@
         </tr>
         <tr class="altColor">
             <td class="colFirst"><a
-                    href="com/github/paolorotolo/appintroexample/util/package-summary.html">com.amqtech.opensource.appintroexample.util</a>
+                    href="com/github/paolorotolo/appintroexample/util/package-summary.html">com.github.appintro.example.util</a>
             </td>
             <td class="colLast">&nbsp;</td>
         </tr>
